change dir structure to add sidebar only to chat pages
