+++ src/app/chats/layout.tsx
@@ -1,0 +1,7 @@
+import { SidebarProvider } from "@components/ui/sidebar";
+
+export default function Home({
+  children,
+}: Readonly<{ children: React.ReactNode }>) {
+  return <SidebarProvider>{children}</SidebarProvider>;
+}
